+++ src/kernel/services/fs/inode.rs
@@ -41,6 +41,9 @@ pub trait Inode: Send + Sync {
     /// - `buf`: 目标缓冲区
     /// - `pwm`: 权限凭证
     /// 返回: 实际读取的字节数
+    ///
+    /// # Errors
+    /// 当偏移超出文件末尾、底层 I/O 失败或权限不足时返回 `KernelError` (具体由实现者决定).
     fn read(&self, offset: u64, buf: &mut [u8], pwm: u64) -> KernelResult<usize>;
 
     /// 写入文件数据
@@ -49,20 +52,32 @@ pub trait Inode: Send + Sync {
     /// - `buf`: 源数据
     /// - `pwm`: 权限凭证
     /// 返回: 实际写入的字节数
+    ///
+    /// # Errors
+    /// 当文件只读、超出文件系统容量或底层 I/O 失败时返回 `KernelError` (具体由实现者决定).
     fn write(&self, offset: u64, buf: &[u8], pwm: u64) -> KernelResult<usize>;
 
     /// 获取文件属性
+    ///
+    /// # Errors
+    /// 当 inode 已失效或底层元数据读取失败时返回 `KernelError` (具体由实现者决定).
     fn stat(&self, pwm: u64) -> KernelResult<VfsStat>;
 
     /// 截断文件到指定大小
+    ///
+    /// # Errors
+    /// 当文件只读、无权限或底层元数据更新失败时返回 `KernelError` (具体由实现者决定).
     fn truncate(&self, size: u64, pwm: u64) -> KernelResult<()>;
 
     /// 计算 seek 后的新偏移
     ///
     /// - `offset`: seek 偏移量 (字节)
-    /// - `whence`: 定位基准 (SEEK_SET/SEEK_CUR/SEEK_END)
+    /// - `whence`: 定位基准 (`SEEK_SET/SEEK_CUR/SEEK_END`)
     /// - `current_offset`: 当前文件偏移
     /// 返回: 新的文件偏移
+    ///
+    /// # Errors
+    /// 当 seek 结果超出可表示范围或 `whence` 非法时返回 `KernelError` (具体由实现者决定).
     fn seek(&self, offset: i64, whence: VfsSeekWhence, current_offset: u64) -> KernelResult<u64>;
 
     /// 是否为目录
@@ -71,66 +86,99 @@ pub trait Inode: Send + Sync {
     /// 读取目录项 (仅目录 Inode 实现)
     ///
     /// - `offset`: 目录偏移 (字节)
-    /// 返回: (entry_name, file_type, has_more), 失败返回 KernelError
+    /// 返回: (`entry_name`, `file_type`, `has_more`), 失败返回 `KernelError`
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotADirectory`; 当 inode 不是目录或底层目录读取失败时返回 `KernelError`.
     fn readdir(&self, _offset: u64) -> KernelResult<(alloc::string::String, VfsFileType, bool)> {
         Err(KernelError::NotADirectory)
     }
 
     /// 创建子目录 (仅目录 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限、名称已存在或底层创建失败时返回 `KernelError`.
     fn mkdir(&self, _name: &str, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
     /// 删除文件 (仅目录 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限、文件不存在或底层删除失败时返回 `KernelError`.
     fn unlink(&self, _name: &str, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
     /// 重命名 (仅目录 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限、源/目标无效或底层重命名失败时返回 `KernelError`.
     fn rename(&self, _old_name: &str, _new_name: &str, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
     /// 创建硬链接 (仅目录 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限、名称已存在或底层链接创建失败时返回 `KernelError`.
     fn link(&self, _name: &str, _target: &dyn Inode, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
     /// 创建符号链接 (仅目录 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限、名称已存在或底层符号链接创建失败时返回 `KernelError`.
     fn symlink(&self, _name: &str, _target: &str, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
     /// 读取符号链接目标 (仅符号链接 Inode 实现)
+    ///
+    /// # Errors
+    /// 默认实现返回 `InvalidArgument`; 当 inode 不是符号链接或目标读取失败时返回 `KernelError`.
     fn readlink(&self, _buf: &mut [u8]) -> KernelResult<usize> {
         Err(KernelError::InvalidArgument)
     }
 
     /// 修改权限
+    ///
+    /// # Errors
+    /// 默认实现恒返回 `Ok(())`; 当无权限或底层元数据更新失败时返回 `KernelError` (具体由实现者决定).
     fn chmod(&self, _mode: u16, _pwm: u64) -> KernelResult<()> {
         Ok(())
     }
 
     /// 修改所有者
+    ///
+    /// # Errors
+    /// 默认实现恒返回 `Ok(())`; 当无权限或底层元数据更新失败时返回 `KernelError` (具体由实现者决定).
     fn chown(&self, _owner_pwm: u64, _group_pwm: u64, _pwm: u64) -> KernelResult<()> {
         Ok(())
     }
 
     /// 设置文件时间戳 (atime, mtime)
     ///
-    /// - `atime`: 访问时间 (纳秒), u64::MAX 表示不修改
-    /// - `mtime`: 修改时间 (纳秒), u64::MAX 表示不修改
+    /// - `atime`: 访问时间 (纳秒), `u64::MAX` 表示不修改
+    /// - `mtime`: 修改时间 (纳秒), `u64::MAX` 表示不修改
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当无权限或底层元数据更新失败时返回 `KernelError` (具体由实现者决定).
     fn set_times(&self, _atime: u64, _mtime: u64, _pwm: u64) -> KernelResult<()> {
         Err(KernelError::NotSupported)
     }
 
-    /// 获取底层 inode_id (供 pcache 等需要 inode 标识的场景使用)
+    /// 获取底层 `inode_id` (供 pcache 等需要 inode 标识的场景使用)
     fn node_id(&self) -> u32;
 
     /// 获取挂载点索引 (供 mmap 等需要挂载点信息的场景使用)
     fn mount_idx(&self) -> u32;
 
-    /// 按 inode_id 直接读取 (mmap prewarm 用, 不经过 OpenFile offset)
+    /// 按 `inode_id` 直接读取 (mmap prewarm 用, 不经过 `OpenFile` offset)
+    ///
+    /// # Errors
+    /// 默认实现返回 `NotSupported`; 当底层读取失败时返回 `KernelError` (具体由实现者决定).
     fn pread_inode(&self, _offset: u64, _buf: &mut [u8], _pwm: u64) -> KernelResult<usize> {
         Err(KernelError::NotSupported)
     }
@@ -192,9 +240,9 @@ impl Inode for AnonymousInode {
     }
 
     fn seek(&self, offset: i64, whence: VfsSeekWhence, current_offset: u64) -> KernelResult<u64> {
-        let file_size = ANONYMOUS_FS
+        let file_size = u64::from(ANONYMOUS_FS
             .get_size(self.inode_id)
-            .unwrap_or(0) as u64;
+            .unwrap_or(0));
         let new_offset = match whence {
             VfsSeekWhence::Set => offset as u64,
             VfsSeekWhence::Cur => current_offset.saturating_add(offset as u64),
@@ -225,9 +273,9 @@ impl Inode for AnonymousInode {
 // RamFs Inode adapter — 将 RamFsData 包装为 Inode trait
 // ============================================================================
 
-/// RamFS 文件 Inode — 将全局 RamFsData 包装为 Inode trait
+/// `RamFS` 文件 Inode — 将全局 `RamFsData` 包装为 Inode trait
 ///
-/// 每个实例对应一个 RamFS 中的文件节点.
+/// 每个实例对应一个 `RamFS` 中的文件节点.
 /// 通过全局 `RAMFS_DATA` 锁实现内部可变性.
 pub struct RamFsInode {
     inode_id: u32,
@@ -235,7 +283,7 @@ pub struct RamFsInode {
 }
 
 impl RamFsInode {
-    /// 创建新的 RamFS Inode
+    /// 创建新的 `RamFS` Inode
     pub fn new(inode_id: u32, mount_idx: u32) -> Self {
         Self { inode_id, mount_idx }
     }
@@ -246,7 +294,7 @@ impl Inode for RamFsInode {
         use crate::kernel::framework::fs::ramfs::ramfs::RAMFS_DATA;
         let mut ramfs = RAMFS_DATA.lock();
         let (bytes_read, _new_offset) = ramfs.read_at_offset(self.inode_id, offset, buf, pwm);
-        if bytes_read == 0 && offset >= ramfs.get_file_size(self.inode_id).unwrap_or(0) as u64 {
+        if bytes_read == 0 && offset >= u64::from(ramfs.get_file_size(self.inode_id).unwrap_or(0)) {
             // EOF
             Ok(0)
         } else if bytes_read > 0 {
@@ -308,7 +356,7 @@ impl Inode for RamFsInode {
     fn seek(&self, offset: i64, whence: VfsSeekWhence, current_offset: u64) -> KernelResult<u64> {
         use crate::kernel::framework::fs::ramfs::ramfs::RAMFS_DATA;
         let ramfs = RAMFS_DATA.lock();
-        let file_size = ramfs.get_file_size(self.inode_id).unwrap_or(0) as u64;
+        let file_size = u64::from(ramfs.get_file_size(self.inode_id).unwrap_or(0));
         let new_offset = match whence {
             VfsSeekWhence::Set => offset as u64,
             VfsSeekWhence::Cur => current_offset.saturating_add(offset as u64),
@@ -356,9 +404,9 @@ impl Inode for RamFsInode {
 // LegacyInode 将 handle + mount_idx 包装为 Inode trait object,
 // 委托给 VFS_MANAGER 中注册的 FileSystem trait 方法.
 
-/// 过渡期 Inode 适配器 — 将 FileSystem 的 opaque handle 包装为 Inode trait
+/// 过渡期 Inode 适配器 — 将 `FileSystem` 的 opaque handle 包装为 Inode trait
 ///
-/// 每次调用 Inode 方法时, 通过 mount_idx 查找 FileSystem trait object,
+/// 每次调用 Inode 方法时, 通过 `mount_idx` 查找 `FileSystem` trait object,
 /// 委托给对应的 `fs_*` 方法. 性能不是最优, 但保证正确性.
 pub struct LegacyInode {
     handle: u32,
@@ -369,7 +417,7 @@ pub struct LegacyInode {
 }
 
 impl LegacyInode {
-    /// 从 FsOpenResult 创建 LegacyInode
+    /// 从 `FsOpenResult` 创建 `LegacyInode`
     pub fn from_fs_result(handle: u32, mount_idx: u32, file_type: u8, rel_path: &str) -> Self {
         Self {
             handle,
@@ -537,12 +585,12 @@ pub fn new_anonymous_inode(inode_id: u32) -> Arc<dyn Inode> {
     Arc::new(AnonymousInode::new(inode_id))
 }
 
-/// 创建 RamFS Inode 的 Arc 包装
+/// 创建 `RamFS` Inode 的 Arc 包装
 pub fn new_ramfs_inode(inode_id: u32, mount_idx: u32) -> Arc<dyn Inode> {
     Arc::new(RamFsInode::new(inode_id, mount_idx))
 }
 
-/// 创建 LegacyInode 的 Arc 包装 (过渡期: 从 FsOpenResult 创建)
+/// 创建 `LegacyInode` 的 Arc 包装 (过渡期: 从 `FsOpenResult` 创建)
 pub fn new_legacy_inode(handle: u32, mount_idx: u32, file_type: u8, rel_path: &str) -> Arc<dyn Inode> {
     Arc::new(LegacyInode::from_fs_result(handle, mount_idx, file_type, rel_path))
 }
